backend/requirements.txt
fastapi==0.116.1
python-multipart==0.0.20
uvicorn==0.35.0
numpy<2
torch==2.1.0
torchvision==0.16.0
Pillow==10.4.0

